refactored plot_layout.py
