+++ python/perf-kernels/tools/plot-layout/blocked/__init__.py
@@ -1,0 +1,3 @@
+from .plot_blocked import generate_blocked_tex
+
+__all__ = ["generate_blocked_tex"]
