+++ python/perf-kernels/tools/plot-layout/blocked/blockedLayout.tex
