+++ python/perf-kernels/tools/plot-layout/dot/__init__.py
@@ -1,0 +1,3 @@
+from .plot_dot import generate_dot_tex
+
+__all__ = ["generate_dot_tex"]
