+++ python/perf-kernels/tools/plot-layout/dot/dotLayout.tex
