+++ python/perf-kernels/tools/plot-layout/lds/__init__.py
@@ -1,0 +1,3 @@
+from .plot_lds import generate_lds_tex
+
+__all__ = ["generate_lds_tex"]
