+++ python/perf-kernels/tools/plot-layout/utils/__init__.py
@@ -1,0 +1,3 @@
+from .utilities import run_bash_command
+
+__all__ = ["run_bash_command"]
