+++ python/perf-kernels/tools/plot-layout/utils/preamble.tex
