+++ python/perf-kernels/tools/plot-layout/utils/utilities.py
@@ -1,0 +1,6 @@
+import subprocess
+
+
+def run_bash_command(commandstring):
+    proc = subprocess.run(commandstring, shell=True, check=True, executable='/bin/bash', stdout=subprocess.PIPE)
+    return proc.stdout.splitlines()
