+++ python/perf-kernels/tools/plot-layout/wmma/__init__.py
@@ -1,0 +1,3 @@
+from .plot_wmma import generate_wmma_tex
+
+__all__ = ["generate_wmma_tex"]
