+++ python/perf-kernels/tools/plot-layout/wmma/wmmaLayout.tex
